Introduce SendLimiter to demarcate between HTTP connections that are actively sending data, and those that are just awaiting a reply.
SendLimiter caps the number that can be actively sending. This is advantageous because the number of active senders can be inferred through testing with some accuracy, and safely held at a static value. While the number awaiting reply depends on the moment by moment throughput of Storage, and so should be allowed to fluctuate.  It can safely be allowed to fluctuate (taking up as few or as many of the remaining go routines as necessary), now that we have SendLimiter capping the number of active senders.