Refactor/source dest mapping for downloads (#158)
* Refactored download code into one re-usable class (parameterizable for multiple remote targets)

* Make disk writes sequential

Avoids the performance issues associated with zero-filling of interim gaps when saving out-of-order.  Also lays groundwork for future MD5 support.

* Put limits on RAM usage for downloads, and pooled memory buffers to reduce CG pressure

* Made download work spread somewhat more evenly over multilpe files

By limiting the amount of "in flight" chunks for invidual files, when we are over half of our target total RAM usage.  We aren't limiting it when we're below half, because sometimes there will be very few files in a job, so limiting the count per file is not universally applicable - it's only only useful when we have lots of files and lots of RAM usage

* Introduce ChunkD and diagnostic logging of status transitions for each chunkID

The logging is for debugging, rather than day-to-day usage

* Control file-write-though behaviour for downloads

For all but very small files, the OS write cache offers us only extra cost (and/or confusion when monitoring performance)