Force body read retries when a downloaded chunk is abnormally slow
Without this, we can see very occasional chunks that are orders of magnitude slower than normal.  That becomes a problem given our desire to process the files sequentially (for MD5 computation and, in some cases, disk performance). Root cause of the slowness is yet to be determined, but this workaround is effective for now.