Fix/enable body read retries and logging (#194)
* Port forced retry and better logging to the azbfs retry reader

(Same as already in Blob/File SDKs)

* Enable logging of retries in response body downloads

* Force body read retries when a downloaded chunk is abnormally slow

Without this, we can see very occasional chunks that are more than an order of magnitude slower than normal.  That becomes a problem given our desire to process the files sequentially (for MD5 computation and, in some cases, disk performance). Root cause of the slowness is yet to be determined, but this workaround is effective for now.

* Update "dep" dependency of blob SDK to latest version