Move ListPaths back to FileSystemClient where it belongs
The recently-updated Swagger was causing AutoRest to generate ListPaths in the Path client, even though its actually a filesystem operation.  In particular, its URL must always point to the file system, using the "directory" parameter to filter to the desired portion of the directory tree.  You can't use the actual URL of the directory - so it doesn't belong on Path client (and, for everything except the root, it didn't work there either).