Constrain RAM usage by the multiSizeSlicePool
Specifically:
Reduce max slice count in medium and large sub pools. No need to reduce it in the really tiny pools, because even at their (unchanged) original max size, they can't consume much RAM.  This reduction of max count in the medium-sized pools is expected to help with the out-of-RAM with small files bug.
And change GC threshold to something more approprinate for our app.  This is also expected to help with the out-of-RAM with small files bug
And gradulally drain unused large pools. This is expected to help with a related but different anticipated case, where there are large files with differing chunk sizes in the job.