Refactor sync filters into comparators
When the destination is traversed second(*), it is impossible to preserve source properties if everything is just done by the destination traverser. (Because it doesn't see source objects).  So in this commit we move the final processing step into a comparator.  Because it can see both objects, it can use the properties of both to correctly schedule the right operation - including preservation of source properties