Fix/too many open files in download (#276)
* Limit number of open files when downloading, to fit within OS's configured limit

Not needed when uploading, because there the number of open files is (more or less) limited by the number of go-routines used to initiate transfers. ("More or less" because  when we do retries for 503s we will tend to open additonal file handles to re-read the retried data).

* Reduce MaxIdleConnsPerHost

To accomodate Linux systems with low open-files-per-user limit (since each open connection counts as a file)

* Log concurrency parameters at start of job

Useful for debugging things like the max concurrent files issue and limits

* Take network connections into account when setting max number of concurrently open files

* Use one HTTP client for the whole job, instead of one for each job part

Previously, when doing one for each job part, we had an issue when the files in the job parts were very small, and there were a lot of files. This meant we got though each job part quickly, so quickly, that its lingering pool of open connections were an issue, especially on resource-constrained systems.  While the connections are eventually cleaned up, they were not cleaned up fast enough to compensate for our fast progression through the job parts.  Now, we use just one HTTP client, so there's just one connection pool for the whole job.

* Add comment to explain max idle conns per host

* Fix test suite to match change in connection handling