Replace separate goroutine for disk read with mutex (#311)
The goroutine approach was slower on very small files (e.g. 10 million files of 50 KB each) to the point that it affected overall throughput.  Presumably because spinning up a goroutine and doing two selects on channels, while cheap in terms of performance, is not as cheap as manipulating an uncontested mutex.