Feature/command line bandwidth cap (#455)
* Rename old pacer-lite to pacedReadSeeker

* Make expectedBytes optional (can be zero)

* Remove use of PacerPolicyFactory for now (since it doesn't do anything)

* Replace old pacer struct with pacer interface

Pacer interface is the existing pacerConsumer, renamed to pacer now that the name conflict with the old pacer is gone

* Add comments about difference between page blob pacing and global pacing

* Tidy pacer method names

* Hook pacer in to app startup (but no command line parameter yet)

* Hook command line parameter into pacer

* Convert Mbps measurements to use base-10 mega, rather than base-2 mebi

Because base 10 is the customary approach for networking.

* Fix up use of contexts in global pacer

* Add extra comment to doc of cap-mbps parameter

* Reduce "wobble" in golbal pacer rate

* Apply global pacing to S2S copies as well

Why not!  E.g. maybe someone has other activity on source or dest account, and so doesn't want us to take up al their throughput allocation.  S2S pacing is, by nature, coarse-grained at the block level (rather than fine grained inside body read/write as for transfers to/from local).  It has to be coarse-grained, because the S2S API is coarse-grained.