Introduce clean cancellation with improved block-blob cleanup behaviour (#576)
Instead of just calling os.Exit on a cancellation, initiate an orderly shutdown. This allows the existing logic to run, which deletes the partially-written files that are affected by the cancellation. (They're in an un-defined state, due to cancellation while they were in-flight).  Previously that logic did not have time to run.

Furthermore, for block blobs, refine the cleanup so that after a cancellation we don't delete existing committed content. (But we do still delete it after _failures_, for consistency with other destinations and to avoid block list problems - e.g. old uncommited blocks with different ID sizes from other tools, or too many uncommited blocks).

This failure behavior (for all destinations including block blobs) remains unchanged from previous releases. It's just the cancellation behavior that has changed.