Tighten up block list access (#639)
Old code was only locking the reference to the slice, not the contents. In theory it doesn't matter, because the code is in the epilogue so nothing else will access it at that time anyway. But this change makes that assumption explicit by making sure we'll be notified if the assumption is ever violated. (with an error dereferenceing the now-nil slice field)