Make main BlobFS pipeline use main xfer log policy
To pick up xfer log improvements for this release. Note that the old BlobFS log was effectively identical to the 10.2.1 version of the xferLogPolicy, so we are not losing anything by doing this.
Old BlobFS log remains used by credential/enumeration pipeline, because we don't have time to re-work that yet.