Allow cancellation during enumeration (#723)
* Correctly respond to Cancelling state when it is entered before job is fully ordered

Works by allowing the transition to cancelled, in this case.  That will then cause application exit. So there is no change to the actual enumeration code, since it gets ended just by the application exit.

* Use a more thread safe approach for tracking ordering of final part

We were getting away with the old code, safely, due to the memory ordering effects of use of channels in the scheduling of the work - but that's not obvious and could, in theory, be broken by future maintenance.