Add FreeBSD support (#188)
As discussed in #768, I renamed credCacheInternal_linux.go to
credCacheGnomeKeyring_unix.go with the intention to provide a generic
libsecret (née GNOME Keyring) credential cache implementation.  Continue to
build this implementation in the (linux && se_integration) target, but
additionally build this implementation for the FreeBSD target, which lacks
an in-kernel credential cache service.

Rename credCacheGnomeKeyring_linux.go to credCacheGnomeKeyringBindings_unix.go
to reflect that this file contains generic cgo wrappers of the GnomeKeyring
(libsecret) API (neither Linux-specific, nor an implementation of the
CredCache interface).  Add wrappers for secret_password_store() (add) and
secret_password_clear() (delete).

The struct CredCacheInternalIntegration was mechanically renamed to
CredCacheGnomeKeyring to more closely match its new use.  Additionally,
trivial implementations of removeCachedTokenInternal() and
saveTokenInternal() were defined for CredCacheGnomeKeyring.

On Linux (credCache_linux), an alias to the old name is defined for the
se_integration target.  On FreeBSD, the type names CredCache and
CredCacheInternalIntegration are both aliases of CredCacheGnomeKeyring.

This means the FreeBSD port pulls in cgo and libsecret.  But it does not add
any new dependency for other platforms.

More trivially, rename mmf_linux to mmf_unix (not Linux-specific) and add
+build rules to enable it on Linux and FreeBSD.  A straightforward
writeThroughFile_freebsd is implemented (thought Go makes this hard on us by
not providing a nice syscall wrapper for fallocate).

I didn't try to run the test suite yet.