Use equality instead of less-than-or-equal when comparing VHD size to target managed disk size (#776)
Because (a) target size is declared when creating the target and (b) it prevents surprises with the completion-time length check later and (c) by checking exactly against the target (which as already had its size validated) this ensures that the source size meets Azures criteria (i.e. an integer number of MiB, plus 512 bytes for VHD footer).