Bug6974627: Ignore Permisson mismatch while listing parent directory
Transfer to a filestore with a read-only SAS fails because we try
listing properties for parent directory, and fail with permission mismatch.
With this change, we ignore the error and try creating parent directories.
Also, we try matching the source and destination file lengths in
epilogue. The error here will be selectively ignored as well.