Fix azcopy on architectures that are strict about alignment (ARM)
The current version of azcopy dies on ARM systems with a SIGBUS and the following stacktrace:

  INFO: Scanning...
  INFO: Any empty folders will not be processed, because source and/or destination doesn't have full folder support
  unexpected fault address 0xffff72640a5f
  fatal error: fault
  [signal SIGBUS: bus error code=0x1 addr=0xffff72640a5f pc=0x733d8c]

  goroutine 25 [running]:
  runtime.throw(0x9de596, 0x5)
          /usr/lib/go-1.16/src/runtime/panic.go:1117 +0x54 fp=0x4000d982f0 sp=0x4000d982c0 pc=0x43914
  runtime.sigpanic()
          /usr/lib/go-1.16/src/runtime/signal_unix.go:731 +0x284 fp=0x4000d98330 sp=0x4000d982f0 pc=0x5b214
  github.com/Azure/azure-storage-azcopy/v10/common.(*TransferStatus).AtomicLoad(...)
          /home/jeremi/github/azure-storage-azcopy-2/common/fe-ste-models.go:690
  github.com/Azure/azure-storage-azcopy/v10/ste.(*JobPartPlanTransfer).TransferStatus(...)
          /home/jeremi/github/azure-storage-azcopy-2/ste/JobPartPlan.go:377
  github.com/Azure/azure-storage-azcopy/v10/ste.(*jobPartMgr).ScheduleTransfers(0x4000adc000, 0xb54930, 0x4000ad2000)
          /home/jeremi/github/azure-storage-azcopy-2/ste/mgr-JobPartMgr.go:396 +0x42c fp=0x4000d99f20 sp=0x4000d98340 pc=0x733d8c
  github.com/Azure/azure-storage-azcopy/v10/ste.(*jobsAdmin).scheduleJobParts(0x4000496000)
          /home/jeremi/github/azure-storage-azcopy-2/ste/JobsAdmin.go:287 +0x44 fp=0x4000d99fd0 sp=0x4000d99f20 pc=0x724ee4
  runtime.goexit()
          /usr/lib/go-1.16/src/runtime/asm_arm64.s:1130 +0x4 fp=0x4000d99fd0 sp=0x4000d99fd0 pc=0x78a44
  created by github.com/Azure/azure-storage-azcopy/v10/ste.initJobsAdmin
          /home/jeremi/github/azure-storage-azcopy-2/ste/JobsAdmin.go:212 +0x5ec

What happens is that a JobPartPlanTransfer struct is created inside a memory
mapped 'plan' file and there are several members of that struct that are
accessed atomically. This requires that those members have the natural
alignment for their type, which is guaranteed at the struct level, but the
struct itself is not properly aligned inside the plan file. JobPartPlanTransfer
is located after the fixed size JobPartPlanHeader struct and a variable size
'CommandString'. It is that variable sized string that breaks the whole
alignment (most of the time). x86 doesn't care but ARM does not like it
one bit and SIGBUS.

To remedy this, ensure that the start of the first JobPartPlanTransfer is 8
byte aligned both when writing the file and when accessing it. This makes
azcopy work on ARM64 systems that I have tested.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>