Fixing a race condition in jobMgr.DeferredCleanupJobMgr
The call to job manager cancel preceds the call to cleanup transfer routine and it causes a unintended jobPartMgr.get which causes a panic as no job is found for the id. In DeferredCleanupJobMgr, jm.cleanupTransferRoutine sends the cancellation through reportCancelCh to report job part completion and jm.Cancel sends an empty jobPartProgressInfo. Because of the order, the empty jobPartProgressInfo causes a panic in reportJobPartDoneHandler. If we cancel the transfer routine properly before cancel, we can avoid this issue without any functionality impact.

This will not manifest in regular customer flow but we will notice this if the  job status is queried periodically like in XDM Mover.