Resetting transfer lists in JobSummary based on a flag
GetJobSummary\ListJobSummary calls resets the list objects during the get call. This is problematic as its serve dual purpose and a caller can accidentally clear the objects prematurely. Added a parameter to conditionally reset the lists so the caller can decide. When build mode is Mover, regular GetJobSummary calls will skip resetting the lists so that Mover client gets the complete data.