BugFix: Readonly attribute is not being set in target
If source object has readonly attribute, we set it to false in prologue so the object can be uploaded to target. Then in epilogue, we set the readonly property as needed. The bug is introduced due to pointer copy. When readonly is set to false in epilogue, it also modifies the original properties object so epilogue also sets it as false.